Initial commit of ViT: Pytorch based for now
`Pytorch_ViT.ipynb` contains dataset - dataloader, data augmentation,
model definition and training loop along with utilities for tracking
useful training statistics and plotting them.
Note that the model definition and data augmentation along with hyper
parameters and learning rate are set and written haphazardly.
There are todos in the code & some ideas I have that I'll introduce in
the next few commits.

The code is written purely by hand (i.e. no LLM was used in the proccess
of development), since it was written during the Israeli war on Iran and
the ensuing internet suppression by the government.